pyproject.toml
[project]
name = "house"
version = "0.1.0"
description = "学区地图数字化与交互查询系统"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11.9",
    "python-docx>=1.0.0",
    "pandas>=2.1.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "tabulate>=0.9.0",
    "volcengine-python-sdk[ark]",
    "paddlepaddle==3.2.0",
    "paddleocr>=3.0.0",
]

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "paddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cpu/"
